bug fix: Amplify failing to fetch `@rollup/rollup-linux-x64-gnu`
The npm ci Command: The command npm ci is designed for continuous integration. It's very strict and installs dependencies exactly as they are listed in package-lock.json. Due to a known bug in some versions of npm, npm ci fails to correctly fetch the optional Linux-specific package (@rollup/rollup-linux-x64-gnu) needed on the server, even though it knows it's on a Linux machine.